OCR/LLM/requirements.txt
# LLM OCR Requirements
# For deep learning and LLM-based OCR with vision models

# OCR and Deep Learning
easyocr==1.7.1
pytesseract==0.3.13

# Deep Learning frameworks
torch==2.4.1
torchvision==0.19.1
transformers==4.44.2

# LLM and Vision Models
ollama==0.3.3
langchain==0.2.16
langchain-community==0.2.16

# Image processing
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4

# HTTP requests
requests==2.32.3

# Utilities
python-dotenv==1.0.1

# Optional: For advanced features
# keras-ocr==0.9.3
# tensorflow==2.17.0
